requirements.txt
alpaca-py
yfinance
python-dotenv
pandas
